requirements.txt
aiohttp
requests
slixmpp
beautifulsoup4
redis
hyper
PyJWT
cryptography